+++ packages/web/src/index.cdn.ts
@@ -1,3 +1,5 @@
+/* eslint-disable no-underscore-dangle */
+
 import 'regenerator-runtime/runtime';
 
 import {
@@ -9,18 +11,32 @@ import {
   MagicExtensionWarning,
   MagicRPCError,
   MagicSDKWarning,
+  isPromiEvent,
 } from '@magic-sdk/provider';
 import * as types from '@magic-sdk/types';
+import localForage from 'localforage';
+import memoryDriver from 'localforage-driver-memory';
 import { IframeController } from './iframe-controller';
 import { WebTransport } from './web-transport';
 
 const Magic = createSDK(SDKBase, {
   target: 'web',
   sdkName: 'magic-sdk',
-  version: process.env.WEB_VERSION!,
+  version: '%WEB_VERSION%',
   defaultEndpoint: 'https://auth.magic.link/',
   ViewController: IframeController,
   PayloadTransport: WebTransport,
+  configureStorage: /* istanbul ignore next */ async () => {
+    const lf = localForage.createInstance({
+      name: 'MagicAuthLocalStorageDB',
+      storeName: 'MagicAuthLocalStorage',
+    });
+
+    await lf.defineDriver(memoryDriver);
+    await lf.setDriver([localForage.INDEXEDDB, localForage.LOCALSTORAGE, memoryDriver._driver]);
+
+    return lf;
+  },
 });
 
 // NOTE: enums are emitted by TypeScript -- in the CDN bundle we attach public
@@ -33,6 +49,7 @@ Object.assign(Magic, {
   RPCError: MagicRPCError,
   SDKWarning: MagicSDKWarning,
   Extension,
+  isPromiEvent,
 });
 
 export { Magic as default };