+++ packages/web/src/index.ts
@@ -1,6 +1,8 @@
-import 'regenerator-runtime/runtime';
+/* eslint-disable no-underscore-dangle */
 
 import { SDKBase, createSDK } from '@magic-sdk/provider';
+import localForage from 'localforage';
+import * as memoryDriver from 'localforage-driver-memory';
 import { IframeController } from './iframe-controller';
 import { WebTransport } from './web-transport';
 
@@ -12,6 +14,8 @@ export {
   MagicRPCError as RPCError,
   MagicSDKWarning as SDKWarning,
   MagicSDKAdditionalConfiguration,
+  PromiEvent,
+  isPromiEvent,
 } from '@magic-sdk/provider';
 
 export * from '@magic-sdk/types';
@@ -19,10 +23,21 @@ export * from '@magic-sdk/types';
 export const Magic = createSDK(SDKBase, {
   target: 'web',
   sdkName: 'magic-sdk',
-  version: process.env.WEB_VERSION!,
+  version: '%WEB_VERSION%',
   defaultEndpoint: 'https://auth.magic.link/',
   ViewController: IframeController,
   PayloadTransport: WebTransport,
+  configureStorage: /* istanbul ignore next */ async () => {
+    const lf = localForage.createInstance({
+      name: 'MagicAuthLocalStorageDB',
+      storeName: 'MagicAuthLocalStorage',
+    });
+
+    await lf.defineDriver(memoryDriver);
+    await lf.setDriver([localForage.INDEXEDDB, localForage.LOCALSTORAGE, memoryDriver._driver]);
+
+    return lf;
+  },
 });
 
 export type Magic = InstanceType<typeof Magic>;