+++ packages/types/CHANGELOG.md
@@ -10,7 +10,38 @@
 
 #### Added
 
-- Added WebAuthn support types
+- ...
+
+## `1.4.5` - 07/22/2020
+
+#### Added
+
+- Export `PromiEvent` type and `isPromiEvent` utility from SDK entry-points (`magic-sdk` and `@magic-sdk/react-native`).
+
+## `1.4.1` through `1.4.4` - 07/13/2020
+
+#### Fixed
+
+- Bug preventing NPM tarball from containing `/dist` files.
+
+## `1.4.0` - 07/13/2020
+
+#### Changed
+
+- Updated build system to use TypeScript project references instead of Microbundle.
+- Pass `targetOrigin` parameter to `postMessage` calls.
+
+## `1.3.1` - 07/08/2020
+
+#### Fixed
+
+- Bug affecting `localforage` type imports causing compilation failure in TypeScript.
+
+## `1.3.0` - 07/08/2020
+
+#### Added
+
+- WebAuthn support.
 
 ## `1.1.0` - 06/23/2020